fix: ajustes no code reviews
