+++ pipelines/datasets/br_rj_isp_estatisticas_seguranca/tasks.py
@@ -1,53 +1,3 @@
-# -*- coding: utf-8 -*-
-"""
-Tasks for br_isp_estatisticas_seguranca
-"""
-
-###############################################################################
-#
-# Aqui é onde devem ser definidas as tasks para os flows do projeto.
-# Cada task representa um passo da pipeline. Não é estritamente necessário
-# tratar todas as exceções que podem ocorrer durante a execução de uma task,
-# mas é recomendável, ainda que não vá implicar em  uma quebra no sistema.
-# Mais informações sobre tasks podem ser encontradas na documentação do
-# Prefect: https://docs.prefect.io/core/concepts/tasks.html
-#
-# De modo a manter consistência na codebase, todo o código escrito passará
-# pelo pylint. Todos os warnings e erros devem ser corrigidos.
-#
-# As tasks devem ser definidas como funções comuns ao Python, com o decorador
-# @task acima. É recomendado inserir type hints para as variáveis.
-#
-# Um exemplo de task é o seguinte:
-#
-# -----------------------------------------------------------------------------
-# from prefect import task
-#
-# @task
-# def my_task(param1: str, param2: int) -> str:
-#     """
-#     My task description.
-#     """
-#     return f'{param1} {param2}'
-# -----------------------------------------------------------------------------
-#
-# Você também pode usar pacotes Python arbitrários, como numpy, pandas, etc.
-#
-# -----------------------------------------------------------------------------
-# from prefect import task
-# import numpy as np
-#
-# @task
-# def my_task(a: np.ndarray, b: np.ndarray) -> str:
-#     """
-#     My task description.
-#     """
-#     return np.add(a, b)
-# -----------------------------------------------------------------------------
-#
-# Abaixo segue um código para exemplificação, que pode ser removido.
-#
-###############################################################################
 import pandas as pd
 import os
 import requests
@@ -58,8 +8,6 @@ from pipelines.utils.utils import (
     log,
 )
 from pipelines.datasets.br_rj_isp_estatisticas_seguranca.utils import (
-    dict_original,
-    dict_arquitetura,
     change_columns_name,
     create_columns_order,
     check_tipo_fase,
@@ -140,11 +88,11 @@ def clean_data(
         log(f"file -> {file_name} read")
 
     # find new df name
-    novo_nome = dict_original()[file_name]
+    novo_nome = isp_constants.dict_original()[file_name]
 
     log("renaming columns")
     # rename columns
-    link_arquitetura = dict_arquitetura()[novo_nome]
+    link_arquitetura = isp_constants.dict_arquitetura()[novo_nome]
     nomes_colunas = change_columns_name(link_arquitetura)
     df.rename(columns=nomes_colunas, inplace=True)
